backend/pyproject.toml
# =============================================================================
# pyproject.toml — Python Project Configuration (PEP 621)
# =============================================================================
# PURPOSE: Define all Python dependencies and project metadata in one place.
# This replaces the older requirements.txt approach with a modern standard
# that pip, setuptools, and other tools understand natively.
#
# Install dependencies:
#   pip install .              → install all dependencies
#   pip install -e .           → install in editable mode (for development)
#   pip install .[dev]         → install with dev/test extras (future)
# =============================================================================

[project]
name = "energy-platform-backend"
version = "0.3.0"
description = "AI Energy Optimization Platform — Backend API for hotels and commercial buildings"
requires-python = ">=3.12"

dependencies = [
    # -------------------------------------------------------------------------
    # WEB FRAMEWORK
    # -------------------------------------------------------------------------
    # FastAPI: Modern async Python web framework with automatic OpenAPI docs.
    # Uvicorn: ASGI server that runs FastAPI (standard extras include uvloop).
    # -------------------------------------------------------------------------
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",

    # -------------------------------------------------------------------------
    # DATABASE
    # -------------------------------------------------------------------------
    # SQLAlchemy: ORM and query builder. The [asyncio] extra enables async support.
    # asyncpg: High-performance async PostgreSQL driver (used by the FastAPI app).
    # psycopg2-binary: Sync PostgreSQL driver (used by Alembic for migrations).
    # Alembic: Database migration tool that tracks schema changes over time.
    # -------------------------------------------------------------------------
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.14.0",

    # -------------------------------------------------------------------------
    # AUTHENTICATION
    # -------------------------------------------------------------------------
    # python-jose: JWT token creation and verification (HS256 signing).
    # bcrypt: Password hashing (industry-standard for user passwords).
    # python-multipart: Required by FastAPI for OAuth2 form-based login parsing.
    # -------------------------------------------------------------------------
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.18",

    # -------------------------------------------------------------------------
    # CONFIGURATION
    # -------------------------------------------------------------------------
    # pydantic-settings: Type-safe configuration loading from .env files and
    # environment variables. Extends Pydantic's validation to app settings.
    # -------------------------------------------------------------------------
    "pydantic-settings>=2.6.0",

    # -------------------------------------------------------------------------
    # UTILITIES
    # -------------------------------------------------------------------------
    # python-dotenv: Loads .env files into environment variables.
    # Used by pydantic-settings under the hood.
    # -------------------------------------------------------------------------
    "python-dotenv>=1.0.1",
    "redis[hiredis]>=5.0.0",

    # -------------------------------------------------------------------------
    # EMAIL (Phase 2 — Notifications)
    # -------------------------------------------------------------------------
    # SendGrid: Transactional email delivery for alert notifications.
    # -------------------------------------------------------------------------
    "sendgrid>=6.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "testcontainers[postgres]>=4.0.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
addopts = "--cov=app --cov-report=term-missing"
